+ some work around removing compile-time warnings / ide suggestions

+++ haikudepotserver-packagefile/src/main/java/org/haikuos/pkg/PkgFactory.java
@@ -65,7 +65,7 @@ public class PkgFactory {
         Integer revision = null;
 
         if(revisionAttribute.isPresent()) {
-            revision = ((BigInteger) ((IntAttribute) revisionAttribute.get()).getValue(attributeContext)).intValue();
+            revision = ((IntAttribute) revisionAttribute.get()).getValue(attributeContext).intValue();
         }
 
         return new PkgVersion(
